lspark-p/requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asn1crypto==1.5.1
bech32==1.2.0
blessed==1.20.0
blinker==1.8.2
cbor2==5.6.4
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
coincurve==19.0.1
cryptography==43.0.1
cytoolz==1.0.0
dnspython==2.7.0
ecdsa==0.19.0
eciespy==0.4.2
enlighten==1.12.4
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==5.0.0
eth-utils==5.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httpx==0.27.0
ic-py==1.0.1
icpp-candid==4.1.0
icpp-pro==4.1.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
leb128==1.0.8
lightspark==2.10.1
lightspark-crypto-python==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mnemonic==0.20
packaging==24.1
pluggy==1.5.0
prefixed==0.8.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pytest==8.3.2
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
toolz==1.0.0
typer==0.12.4
typing_extensions==4.12.2
uma-sdk==1.2.2
urllib3==2.2.2
waiter==1.5
wcwidth==0.2.13
Werkzeug==3.0.4
zstandard==0.23.0
